CRM/CRMBusinessLogicLayer/CustomerRepository.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace CRMBusinessLogicLayer
{
    public class CustomerRepository
    {
        public Customer Retrieve(int customerId)
        {
            Customer customer = new Customer(customerId);
            if (customerId == 1)
            {
                customer.EmailAddress = "mano@gmail.com";
                customer.FirstName = "Jonas";
                customer.LastName = "Baggins";
            }
            return customer;
        }
    }
}
